Refactor UI component out of AbstractToolBox.
If it belongs in the base class ToolBox which is more concerned with Galaxy-spcifics like UI, AbstractToolBox is suppose to be an abstract description of a toolbox - it is about data management.
